split retrofit
