+++ app/src/main/java/ru/tagilov/avitotrainee/city/di/CityComponent.kt
@@ -14,7 +14,6 @@ import javax.inject.Scope
 interface CityComponent {
 
     @CityScreenScope
-    @FlowPreview
     fun getViewModel(): CityViewModel
 
     @Component.Factory
