+++ app/src/main/java/ru/tagilov/avitotrainee/forecast/di/ForecastModule.kt
@@ -2,13 +2,26 @@ package ru.tagilov.avitotrainee.forecast.di
 
 import dagger.Binds
 import dagger.Module
-import ru.tagilov.avitotrainee.forecast.data.ForecastRepository
-import ru.tagilov.avitotrainee.forecast.data.ForecastRepositoryImpl
-import ru.tagilov.avitotrainee.forecast.data.LocationRepository
-import ru.tagilov.avitotrainee.forecast.data.LocationRepositoryImpl
+import dagger.Provides
+import retrofit2.Retrofit
+import retrofit2.create
+import ru.tagilov.avitotrainee.forecast.data.*
+import ru.tagilov.avitotrainee.root.di.Forecast
+import ru.tagilov.avitotrainee.root.di.Location
 
 @Module(includes = [ForecastBinds::class])
-class ForecastModule
+class ForecastModule {
+    @Provides
+    fun provideForecastApi(
+            @Forecast forecastRetrofit: Retrofit
+    ) : ForecastApi = forecastRetrofit.create()
+
+    @Provides
+    fun provideLocationApi(
+            @Location locationRetrofit: Retrofit
+    ) : LocationApi = locationRetrofit.create()
+
+}
 
 @Module
 interface ForecastBinds {
