+++ app/src/main/java/ru/tagilov/avitotrainee/root/di/Dependencies.kt
@@ -1,5 +1,6 @@
 package ru.tagilov.avitotrainee.root.di
 
+import retrofit2.Retrofit
 import ru.tagilov.avitotrainee.city.data.CityApi
 import ru.tagilov.avitotrainee.core.db.AppDatabase
 import ru.tagilov.avitotrainee.forecast.data.ForecastApi
@@ -9,12 +10,11 @@ interface BaseDependencies {
     val database: AppDatabase
 }
 
-
 interface CityDependencies: BaseDependencies {
-    val cityApi: CityApi
+    @get:Forecast val forecastRetrofit: Retrofit
 }
 
 interface ForecastDependencies: BaseDependencies {
-    val forecastApi: ForecastApi
-    val locationApi: LocationApi
+    @get:Forecast val forecastRetrofit: Retrofit
+    @get:Location val locationRetrofit: Retrofit
 }
